Revert "Switch to new price measurements"

+++ lib/sanbase_web/graphql/schema/project_types.ex
@@ -163,15 +163,15 @@ defmodule SanbaseWeb.Graphql.ProjectTypes do
       resolve(&ProjectResolver.percent_change_7d/3)
     end
 
-    field :funds_raised_usd_ico_end_price, :float do
+    field :funds_raised_usd_ico_end_price, :decimal do
       cache_resolve(&ProjectResolver.funds_raised_usd_ico_end_price/3)
     end
 
-    field :funds_raised_eth_ico_end_price, :float do
+    field :funds_raised_eth_ico_end_price, :decimal do
       cache_resolve(&ProjectResolver.funds_raised_eth_ico_end_price/3)
     end
 
-    field :funds_raised_btc_ico_end_price, :float do
+    field :funds_raised_btc_ico_end_price, :decimal do
       cache_resolve(&ProjectResolver.funds_raised_btc_ico_end_price/3)
     end
 
@@ -255,15 +255,15 @@ defmodule SanbaseWeb.Graphql.ProjectTypes do
     field(:tokens_issued_at_ico, :decimal)
     field(:tokens_sold_at_ico, :decimal)
 
-    field :funds_raised_usd_ico_end_price, :float do
+    field :funds_raised_usd_ico_end_price, :decimal do
       resolve(&IcoResolver.funds_raised_usd_ico_end_price/3)
     end
 
-    field :funds_raised_eth_ico_end_price, :float do
+    field :funds_raised_eth_ico_end_price, :decimal do
       resolve(&IcoResolver.funds_raised_eth_ico_end_price/3)
     end
 
-    field :funds_raised_btc_ico_end_price, :float do
+    field :funds_raised_btc_ico_end_price, :decimal do
       resolve(&IcoResolver.funds_raised_btc_ico_end_price/3)
     end
